Rename module wrfuzzer to wrfuzzer and split its service functionality and split its arg parser into modular components/services
