--- bin/wffuzzer
@@ -1,14 +1,0 @@
-#!/usr/bin/python
-
-# author: WatchDogOblivion
-# description: TODO
-# WatchDogs File Fuzzer a.k.a wffuzzer
-
-from watchdogs.web.scripts import FileFuzzerScript
-
-
-def __main__():  #type: () -> None
-  FileFuzzerScript().run()
-
-
-__main__()
