+++ watchdogs/web/models/__init__.py
@@ -1,1 +1,3 @@
-from FileFuzzer import FileFuzzer
+from .RequestFuzzer import RequestFuzzer
+from .Response import Response
+from .WebFile import WebFile
