+++ watchdogs/web/parsers/__init__.py
@@ -1,1 +1,2 @@
-from .FileFuzzerArgs import FileFuzzerArgs
+from .RequestArgs import RequestArgs
+from .FuzzerArgs import FuzzerArgs
