+++ watchdogs/web/scripts/__init__.py
@@ -1,1 +1,1 @@
-from .FileFuzzerScript import FileFuzzerScript
+from .RequestFuzzerScript import RequestFuzzerScript
