+++ watchdogs/web/services/__init__.py
@@ -1,1 +1,3 @@
-from .FileFuzzerService import FileFuzzerService
+from .RequestParserService import RequestParserService
+from .RequestResponseService import RequestResponseService
+from .RequestResponseFuzzerService import RequestResponseFuzzerService
